Improve mounting hole accuracy
Turns out these holes were actually 26.565 degrees (atan(0.5)), not 22.5.
This fixes the footprint to angle the pad at 26.5 and puts the courtyard
and user comments figures in the exactly correct spot.

This also renames the footprint, switching from specifying the angle
(which would be inexact) to the center offset of the second hole.