Make mounting holes non-plated
It seems that having a pad around an oval hole confuses Aisler and
produces a weird result (at least in the preview, Aisler has not yet
finished investigating whether there is also a problem in actual
production).

As a workaround, make the holes nonplated (and make the pad size equal
to the hole size), since we do not really need the grounding currently
(but this makes minimal changes to allow reverting this later on). This
also sets the "zone connection" to "none", to prevent thermal spokes to
the non-existing pad (without having to modify the schematic to not
connect these pads to GND, which would be the proper way for a regular
NPTH).